+++ stories/SQFormAsyncAutocomplete.stories.js
@@ -57,12 +57,13 @@ const defaultArgs = {
 };
 
 const Template = args => {
-  const {schema, ...rest} = args;
+  const {schema, SQFormProps, ...rest} = args;
   return (
     <div style={{minWidth: 250}}>
       <SQFormStoryWrapper
         initialValues={{[defaultArgs.name]: ''}}
         validationSchema={schema}
+        {...SQFormProps}
       >
         <SQFormAsyncAutocomplete
           {...rest}